requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2023.11.17
charset-normalizer==3.4.0
click==8.1.8
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3